requirements.txt
feedparser==6.0.8
pandas==1.3.3
pyexasol==0.20.0
python==3.7.0
python_dateutil==2.8.2
requests==2.26.0
streamlit==1.9.0
